Casting to double, so avg precision increases
git-svn-id: svn://hinux.hin.no/home/DT2006/murray/svn/gh@226 c6f41995-d01f-4a11-924a-98aa25776321

+++ trunk/grouphug/modules/WordCount.java
@@ -99,7 +99,7 @@ public class WordCount implements GrouphugModule {
                 Object[] values = sql.getValueList();
                 long words = ((Long)values[2]);
                 long lines = ((Long)values[3]);
-                long wpl = words / lines;
+                double wpl = (double)words / (double)lines;
                 reply += (place++)+". "+ values[1]+ " ("+words+" words, "+lines+" lines, "+
                         (new DecimalFormat("0.0")).format(wpl)+
                         " wpl)\n";
@@ -131,7 +131,7 @@ public class WordCount implements GrouphugModule {
                 Object[] values = sql.getValueList();
                 long words = ((Long)values[2]);
                 long lines = ((Long)values[3]);
-                long wpl = words / lines;
+                double wpl = (double)words / (double)lines;
 
                 bot.sendMessage(nick + " has uttered "+words+ " words in "+lines+" lines ("+
                         (new DecimalFormat("0.0")).format(wpl)+
